fix(cockpit/admin): Migrate App Pages from northware-admin to northware-cockpit
