--- apps/northware-admin/src/app/login/[[...login]]/page.tsx
@@ -1,7 +1,0 @@
-import { LoginForm } from "@northware/ui/components/login";
-
-const LoginPage = () => {
-  return <LoginForm service="admin" />;
-};
-
-export default LoginPage;
