+++ apps/northware-cockpit/src/components/create-user-form.tsx
@@ -16,7 +16,6 @@ import {
 } from "@northware/ui/components/form";
 import { Input } from "@northware/ui/components/input";
 import { useRouter } from "next/navigation";
-
 import { useState } from "react";
 import { type SubmitHandler, useForm } from "react-hook-form";
