+++ apps/northware-cockpit/src/lib/user-actions.ts
@@ -1,7 +1,12 @@
 "use server";
 
-import type { TCreateUserFormSchema } from "@/lib/user-schema";
-import { clerkClient } from "@northware/auth/server";
+import type {
+  TChangePasswordFormSchema,
+  TCreateEMailAddressFormSchema,
+  TCreateUserFormSchema,
+  TUpdateUserFormSchema,
+} from "@/lib/user-schema";
+import { clerkClient, currentUser } from "@northware/auth/server";
 import { db } from "@northware/database/connection";
 import { rolesTable, rolesToAccounts } from "@northware/database/schema";
 import { and, eq, inArray } from "drizzle-orm";
@@ -16,12 +21,122 @@ interface ClerkError {
   }>;
 }
 
+function handleClerkError(typesafeError: ClerkError) {
+  const errorMessages: string[] = [];
+  if (typesafeError.errors) {
+    typesafeError.errors.map((error) => {
+      switch (error.code) {
+        case "form_password_length_too_short":
+          errorMessages.push(
+            "Das Passwort muss mindestens 8 Zeichen lang sein."
+          );
+          break;
+        case "form_password_size_in_bytes_exceeded":
+          errorMessages.push("Das Passwort darf maximal 72 Zeichen lang sein.");
+          break;
+        case "form_password_pwned":
+          errorMessages.push(
+            "Dies ist ein kompromitiertes Passwort. Bitte nutzen Sie ein anderes Passwort."
+          );
+          break;
+        case "form_username_invalid_length":
+          errorMessages.push(
+            "Der Benutzername muss zwischen 4 und 64 Zeichen lang sein."
+          );
+          break;
+        case "form_param_format_invalid":
+          errorMessages.push("Bitte geben Sie eine gültige E-Mail Adresse an.");
+          break;
+        case "form_identifier_exists":
+          switch (error.meta.paramName) {
+            case "email_address":
+              errorMessages.push(
+                "Diese E-Mail Adresse ist bereits registriert."
+              );
+              break;
+            case "username":
+              errorMessages.push("Dieser Benutzername ist bereits vergeben.");
+              break;
+            default:
+              errorMessages.push(
+                "Der Username oder die E-Mail-Adresse ist bereits vergeben."
+              );
+              break;
+          }
+          break;
+        default:
+          errorMessages.push(
+            `Es ist ein Fehler aufgetreten: ${error.message} (Fehlercode: ${error.code})` ||
+              "Es ist ein unbekannter Fehler bei der Kommunikation mit dem Server aufgetreten."
+          );
+          break;
+      }
+    });
+  } else {
+    errorMessages.push(
+      "Es ist ein unbekannter Fehler beim Aktualisieren des Nutzers aufgetreten."
+    );
+  }
+  if (errorMessages.length > 0) {
+    throw new Error(JSON.stringify(errorMessages));
+  }
+}
+
 export async function getUsers() {
   try {
+    const loggedInUser = await currentUser();
     const client = await clerkClient();
     const response = await client.users.getUserList();
-    return response;
+    const users = response.data
+      .filter((user) => user.id !== loggedInUser?.id)
+      .map((user) => {
+        const primaryEmail = user.emailAddresses.find(
+          (email) => email.id === user.primaryEmailAddressId
+        )?.emailAddress;
+        return {
+          id: user.id,
+          fullName: user.fullName,
+          email: primaryEmail,
+          username: user.username,
+        };
+      });
+    return users;
+  } catch {
+    return null;
+  }
+}
+
+export async function getSingleUser(id: string) {
+  try {
+    const client = await clerkClient();
+    const response = await client.users.getUser(id);
+    const user_emailAddresses = response.emailAddresses
+      .map((email) => ({
+        id: email.id,
+        emailAddress: email.emailAddress,
+        verificationStatus: email.verification?.status,
+      }))
+      .sort((a, b) => {
+        if (a.id === response.primaryEmailAddressId) {
+          return -1; // 'a' kommt zuerst
+        }
+        if (b.id === response.primaryEmailAddressId) {
+          return 1; // 'b' kommt zuerst
+        }
+        // Alphabetische Sortierung der übrigen E-Mail-Adressen
+        return a.emailAddress.localeCompare(b.emailAddress);
+      });
+    return {
+      id: response.id,
+      firstName: response.firstName,
+      lastName: response.lastName,
+      fullName: response.fullName,
+      username: response.username,
+      emailAddresses: user_emailAddresses,
+      primaryEmailAddressId: response.primaryEmailAddressId,
+    };
   } catch (error) {
+    // FIXME: Kann dieser Error in eine global-error Seite eingebaut werden?
     console.error(error);
   }
 }
@@ -39,90 +154,129 @@ export async function createUser(formData: TCreateUserFormSchema) {
     });
     revalidatePath("/admin");
   } catch (error) {
-    const errorMessages: string[] = [];
     const typesafeError = error as ClerkError;
-    if (typesafeError.errors) {
-      typesafeError.errors.map((error) => {
-        switch (error.code) {
-          case "form_password_length_too_short":
-            errorMessages.push(
-              "Das Passwort muss mindestens 8 Zeichen lang sein."
-            );
-            break;
-          case "form_password_size_in_bytes_exceeded":
-            errorMessages.push(
-              "Das Passwort darf maximal 72 Zeichen lang sein."
-            );
-            break;
-          case "form_password_pwned":
-            errorMessages.push(
-              "Dies ist ein kompromitiertes Passwort. Bitte nutzen Sie ein anderes Passwort."
-            );
-            break;
-          case "form_username_invalid_length":
-            errorMessages.push(
-              "Der Benutzername muss zwischen 4 und 64 Zeichen lang sein."
-            );
-            break;
-          case "form_param_format_invalid":
-            errorMessages.push(
-              "Bitte geben Sie eine gültige E-Mail Adresse an."
-            );
-            break;
-          case "form_identifier_exists":
-            switch (error.meta.paramName) {
-              case "email_address":
-                errorMessages.push(
-                  "Diese E-Mail Adresse ist bereits registriert."
-                );
-                break;
-              case "username":
-                errorMessages.push("Dieser Benutzername ist bereits vergeben.");
-                break;
-              default:
-                errorMessages.push(
-                  "Der Username oder die E-Mail-Adresse ist bereits vergeben."
-                );
-                break;
-            }
-            break;
-          default:
-            errorMessages.push(
-              `Es ist ein Fehler aufgetreten: ${error.message} (Fehlercode: ${error.code})` ||
-                "Es ist ein unbekannter Fehler bei der Kommunikation mit dem Server aufgetreten."
-            );
-            break;
-        }
-      });
-    } else {
-      errorMessages.push(
-        "Es ist ein unbekannter Fehler beim Anlegen des Nutzers aufgetreten."
-      );
-    }
-    if (errorMessages.length > 0) {
-      throw new Error(JSON.stringify(errorMessages));
-    }
+    handleClerkError(typesafeError);
   }
 }
 
-export async function getSingleUser(id: string) {
-  try {
-    const client = await clerkClient();
-    const response = await client.users.getUser(id);
-    return response;
-  } catch (error) {
-    console.error(error);
+export async function updateUser(formData: TUpdateUserFormSchema, id?: string) {
+  if (typeof id === "undefined") {
+    new Error(
+      "Beim Aufruf der Funktion wurde nicht angegeben, welche Id der User hat."
+    );
+  } else {
+    const { username, firstName, lastName } = formData;
+
+    try {
+      const client = await clerkClient();
+      await client.users.updateUser(id, {
+        username: username,
+        firstName: firstName,
+        lastName: lastName,
+      });
+      revalidatePath("/user");
+    } catch (error) {
+      const typesafeError = error as ClerkError;
+      handleClerkError(typesafeError);
+    }
   }
 }
 
 export async function deleteUser(id: string) {
   try {
     const client = await clerkClient();
-    const response = await client.users.deleteUser(id);
-
-    console.info(response);
+    await client.users.deleteUser(id);
+    revalidatePath("/user");
   } catch (error) {
-    console.error(error);
+    const typesafeError = error as ClerkError;
+    handleClerkError(typesafeError);
+  }
+}
+
+export async function createEmailAddress(
+  formData: TCreateEMailAddressFormSchema,
+  userId?: string
+) {
+  if (typeof userId === "undefined") {
+    new Error(
+      "Beim Aufruf der Funktion wurde nicht angegeben, welche Id der User hat."
+    );
+  } else {
+    const { emailAddress, verified, primary } = formData;
+    try {
+      const client = await clerkClient();
+      await client.emailAddresses.createEmailAddress({
+        userId: userId,
+        emailAddress: emailAddress,
+        verified: verified,
+        primary: primary,
+      });
+      revalidatePath("/user");
+    } catch (error) {
+      const typesafeError = error as ClerkError;
+      handleClerkError(typesafeError);
+    }
+  }
+}
+
+export async function updateEmailAddress(
+  addressId: string,
+  mode: "primary" | "verification",
+  verification?: boolean
+) {
+  try {
+    const client = await clerkClient();
+    if (mode === "primary") {
+      await client.emailAddresses.updateEmailAddress(addressId, {
+        primary: true,
+      });
+    }
+    if (mode === "verification") {
+      await client.emailAddresses.updateEmailAddress(addressId, {
+        verified: verification,
+      });
+    }
+    revalidatePath("/user");
+  } catch (error) {
+    const typesafeError = error as ClerkError;
+    handleClerkError(typesafeError);
+  }
+}
+
+export async function deleteEmailAddress(addressId: string) {
+  try {
+    const client = await clerkClient();
+    await client.emailAddresses.deleteEmailAddress(addressId);
+    revalidatePath("/user");
+  } catch (error) {
+    const typesafeError = error as ClerkError;
+    handleClerkError(typesafeError);
+  }
+}
+
+export async function changePassword(
+  id: string | undefined,
+  formData: TChangePasswordFormSchema
+) {
+  if (typeof id === "undefined") {
+    new Error(
+      "Beim Aufruf der Funktion wurde nicht angegeben, welche ID der User hat."
+    );
+  } else if (formData.newPassword === formData.confirmPassword) {
+    try {
+      const client = await clerkClient();
+      await client.users.updateUser(id, {
+        password: formData.newPassword,
+        skipPasswordChecks: formData.skipChecks,
+        signOutOfOtherSessions: formData.signOutSessions,
+      });
+      revalidatePath("/user");
+    } catch (error) {
+      const typesafeError = error as ClerkError;
+      handleClerkError(typesafeError);
+    }
+  } else {
+    new Error("Die eingegebenen Passwörter stimmen nicht überein.");
   }
 }
